requirements.txt
flask
beautifulsoup4
requests
lxml
python-dotenv
aria2p
